+++ addOns/openapi/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Correct import of v1.2 definitions (Issue 5262).
 - Fix exception when reporting errors.
 - Update minimum ZAP version to 2.8.0.
+- Add import menu to (new) top level Import menu instead of Tools menu.
 
 ## 12 - 2018-05-18