+++ addOns/soap/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Change default accelerator for "Import a WSDL file from local file system".
 - Fix exception with Java 9+ (Issue 4037).
 - Update minimum ZAP version to 2.8.0.
+- Add import menus to (new) top level Import menu instead of Tools menu.
 
 ## 3 - 2017-03-31